Always export deppack methods. Closes GH-16
Currently, deppack instance methods are only available after it's
initialized, which means Brunch can't have
`const explore = require('deppack').exploreDeps;`

With this change, it will be possible to only import the functions you
need even before deppack itself is initialized.

Since deppack is singleton, i.e., there is not going to be more than one
instance of it, this approach should be okay.